+++ apps/namadillo/src/App/Ibc/ShieldAll.tsx
@@ -1,0 +1,14 @@
+import { PageWithSidebar } from "App/Common/PageWithSidebar";
+
+export const ShieldAll: React.FC = () => {
+  return (
+    <PageWithSidebar>
+      <div className="flex flex-col gap-2 text-yellow">
+        <div>IBC: ShieldAll (WIP)</div>
+      </div>
+      <aside className="flex flex-col gap-2 text-yellow">
+        <div>Sidebar (WIP)</div>
+      </aside>
+    </PageWithSidebar>
+  );
+};
