+++ apps/namadillo/src/App/Ibc/Withdraw.tsx
@@ -1,10 +1,10 @@
 import { PageWithSidebar } from "App/Common/PageWithSidebar";
 
-export const ShieldAll: React.FC = () => {
+export const Withdraw: React.FC = () => {
   return (
     <PageWithSidebar>
       <div className="flex flex-col gap-2 text-yellow">
-        <div>Shield All (WIP)</div>
+        <div>IBC: Withdraw (WIP)</div>
       </div>
       <aside className="flex flex-col gap-2 text-yellow">
         <div>Sidebar (WIP)</div>
