+++ apps/namadillo/src/App/Masp/ShieldedFungibleTable.tsx
@@ -3,10 +3,12 @@ import { formatPercentage } from "@namada/utils";
 import { AtomErrorBoundary } from "App/Common/AtomErrorBoundary";
 import { NamCurrency } from "App/Common/NamCurrency";
 import { TableWithPaginator } from "App/Common/TableWithPaginator";
+import { IbcRoutes } from "App/Ibc/routes";
 import BigNumber from "bignumber.js";
 import { AtomWithQueryResult } from "jotai-tanstack-query";
 import { useEffect, useState } from "react";
 import { twMerge } from "tailwind-merge";
+import { MaspRoutes } from "./routes";
 
 type MockedToken = {
   name: string;
@@ -84,6 +86,11 @@ export const ShieldedFungibleTable = (): JSX.Element => {
           size="xs"
           outlineColor="white"
           className="w-fit mx-auto"
+          href={
+            token.name === "NAM" ?
+              MaspRoutes.unshield().url
+            : IbcRoutes.withdraw().url
+          }
         >
           Unshield
         </ActionButton>,
@@ -116,14 +123,7 @@ export const ShieldedFungibleTable = (): JSX.Element => {
       niceError="Unable to load your validators list"
       containerProps={{ className: "pb-16" }}
     >
-      <ActionButton
-        size="xs"
-        outlineColor="white"
-        className="w-fit ml-auto mt-6"
-      >
-        Unshield ALL
-      </ActionButton>
-      <div className="text-sm font-medium">
+      <div className="text-sm font-medium mt-6">
         <span className="text-yellow">{list.length} </span>
         Tokens
       </div>
