+++ apps/namadillo/src/App/Transfer/Transfer.tsx
@@ -2,7 +2,6 @@ import { NotFound } from "App/Common/NotFound";
 import { Route, Routes } from "react-router-dom";
 import { Example } from "./Example";
 import { NamTransfer } from "./NamTransfer";
-import { ShieldAll } from "./ShieldAll";
 import { TransferRoutes } from "./routes";
 
 export const Transfer: React.FC = () => (
@@ -10,10 +9,6 @@ export const Transfer: React.FC = () => (
     <Route
       path={TransferRoutes.namTransfer().toString()}
       element={<NamTransfer />}
-    />
-    <Route
-      path={TransferRoutes.shieldAll().toString()}
-      element={<ShieldAll />}
     />
     <Route path={TransferRoutes.example().toString()} element={<Example />} />
     <Route path="*" element={<NotFound />} />
