+++ apps/namadillo/src/App/Transfer/routes.ts
@@ -8,14 +8,11 @@ export const overview = (): RouteOutput => routeOutput("/");
 
 export const namTransfer = (): RouteOutput => routeOutput("/nam");
 
-export const shieldAll = (): RouteOutput => routeOutput(`/shield-all`);
-
 export const example = (): RouteOutput => routeOutput(`/example`);
 
 export const TransferRoutes = {
   index,
   overview,
   namTransfer,
-  shieldAll,
   example,
 };
